# treasure-planet/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/getfriends": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Получить список друзей пользователя",
                "description": "Возвращает список друзей игрока со статусом 'accepted'.",
                "operationId": "e63830015fa33be157c71a6c084cad1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Список друзей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "players": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Player"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/getplayers": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Получить список пользователей",
                "description": "Возвращает список игроков.",
                "operationId": "c6cb2b84586f608788accde4b8da5b0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Список игроков",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "players": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Player"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/getfriendsinvites": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Получить список входящих заявок в друзья",
                "description": "Возвращает список заявок со статусом 'pending'.",
                "operationId": "dd4785937ee3b216dc9841d93b2ab7de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Список заявок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "friends": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FriendRequest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sendfriendinvite": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Отправить заявку в друзья",
                "description": "Создаёт заявку в друзья от одного игрока к другому.",
                "operationId": "3e5550aa97e7400322eb5dc3725653bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "friend_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заявка отправлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Заявка отправлена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка отправки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка отправки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/acceptfriendinvite": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Принять заявку в друзья",
                "description": "Принимает входящую заявку и устанавливает статус 'accepted'.",
                "operationId": "3517ea2be67549e308bea974d00aab50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "friend_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Друг добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Друг добавлен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка добавления",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка добавления"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/declinefriendinvite": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Отклонить заявку в друзья",
                "description": "отклоняет входящую заявку и удаляет её.",
                "operationId": "cdaba52793d0ed89888be6bf2f7e4a45",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "friend_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заявка удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Друг добавлен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка удаления",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка добавления"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deletefriend": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Удалить друга",
                "description": "Удаляет друга из списка друзей.",
                "operationId": "656e4641fb1841cd3471e53667373903",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "friend_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Произошло удаление",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Произошло удаление"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка удаления",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ошибка удаления"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Player": {
                "properties": {
                    "id": {
                        "description": "ID игрока",
                        "type": "integer"
                    },
                    "username": {
                        "description": "Имя игрока",
                        "type": "string"
                    },
                    "avatar_id": {
                        "description": "Avatar игрока",
                        "type": "integer"
                    },
                    "hp": {
                        "description": "HP игрока",
                        "type": "integer"
                    },
                    "money": {
                        "description": "Money игрока",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "FriendRequest": {
                "properties": {
                    "player_id": {
                        "description": "ID игрока, который отправил заявку",
                        "type": "integer"
                    },
                    "friend_id": {
                        "description": "ID игрока, которому отправлена заявка",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Статус заявки (pending, accepted)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}